[FIX][BP] Use _inherit=base instead of monkeypatching
