+++ multisearch_field/__init__.py
@@ -1,1 +1,1 @@
-from . import monkeypatching
+from . import models
